cleaning the frontend removing old
